git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/ERM.BL/Stable/BLFlex;C24349

+++ Source/ApplicationServices/2Gis.Erm.BLFlex.Operations.Global/Shared/Formatter/RussianWordPluralizer.cs
@@ -15,7 +15,7 @@ namespace DoubleGis.Erm.BLFlex.Operations.Global.Shared.Formatter
             _currencyNameForFiveAndMore = currencyNameForFiveAndMore;
         }
 
-        public string GetPluralFor(int numberToRead)
+        public string GetPluralFor(long numberToRead)
         {
             numberToRead = Math.Abs(numberToRead);
